move logic from chart-type/dimension to series-chart
