--- packages/core/addon/templates/components/visualization-config/chart-type/dimension.hbs
@@ -1,9 +1,0 @@
-{{!-- Copyright 2019, Yahoo Holdings Inc. Licensed under the terms of the MIT license. See accompanying LICENSE.md file for terms. --}}
-{{chart-series-collection
-    allSeriesData=allSeriesData
-    selectedSeriesData=selectedSeriesData
-    seriesDimensions=dimensions
-    maxSeries=maxSeries
-    onRemoveSeries=(action "onRemoveSeries")
-    onAddSeries=(action "onAddSeries")
-}}
