--- packages/core/addon/templates/components/visualization-config/chart-type/metric.hbs
@@ -1,1 +1,0 @@
-{{!-- Copyright 2019, Yahoo Holdings Inc. Licensed under the terms of the MIT license. See accompanying LICENSE.md file for terms. --}}
