+++ packages/core/addon/templates/components/visualization-config/line-chart.hbs
@@ -1,58 +1,60 @@
 {{!-- Copyright 2019, Yahoo Holdings Inc. Licensed under the terms of the MIT license. See accompanying LICENSE.md file for terms. --}}
-{{#if (or (eq type "line-chart") showStackOption)}}
-  <span class="line-chart-config__header">Chart Style Options</span>
-  <div class="line-chart-config__style-options">
-    {{#if (eq type "line-chart")}}
-      <div class="line-chart-config__curve-opt">
-        <span class="line-chart-config__curve-opt-title">Line Style</span>
-        {{#power-select
-          triggerClass="line-chart-config__curve-opt-select"
-          options=curveOptions
-          selected=(readonly options.style.curve)
-          onchange=(action "onUpdateStyle" "curve")
-          searchEnabled=false
-          placeholder="Line"
-          as |curve|}}
-          {{capitalize curve}}
-        {{/power-select}}
-      </div>
-      <div class="line-chart-config__area-opt">
-        <span class="line-chart-config__area-opt-title">Area</span>
-        {{x-toggle
-          class="line-chart-config__area-opt-select"
-          value=(readonly options.style.area)
-          size="small"
-          onToggle=(action "onUpdateStyle" "area")
-        }}
-      </div>
-    {{/if}}
-    {{#if showStackOption}}
-      <div class="line-chart-config__stacked-opt">
-        <span class="line-chart-config__stacked-opt-title">Stacked</span>
-        {{x-toggle
-          class="line-chart-config__stacked-opt-select"
-          value=(readonly options.style.stacked)
-          size="small"
-          onToggle=(action "onUpdateStyle" "stacked")
-        }}
-      </div>
-    {{/if}}
-  </div>
-  {{#if (and showStackOption options.style.stacked)}}
-    {{chart-series-config
-      seriesConfig=seriesConfig
-      seriesType=seriesType
-      onUpdateConfig=(action "onUpdateConfig")
-      class="line-chart-config__series-config"
-    }}
-  {{/if}}
-{{else if (not (eq seriesType "dimension"))}}
-  <div class="visualization-config__warn-msg">No configuration options available.</div>
-{{/if}}
-{{component
-  (concat typePrefix seriesType)
+{{#component
+  "visualization-config/series-chart"
   request=request
   response=response
   seriesConfig=seriesConfig
+  seriesType=seriesType
   onUpdateConfig=(action "onUpdateConfig")
 }}
+  {{#if (or (eq type "line-chart") showStackOption)}}
+    <div class="line-chart-config__header">Chart Style Options</div>
+    <div class="line-chart-config__style-options">
+      {{#if (eq type "line-chart")}}
+        <div class="line-chart-config__curve-opt">
+          <span class="line-chart-config__curve-opt-title">Line Style</span>
+          {{#power-select
+            triggerClass="line-chart-config__curve-opt-select"
+            options=curveOptions
+            selected=(readonly options.style.curve)
+            onchange=(action "onUpdateStyle" "curve")
+            searchEnabled=false
+            placeholder="Line"
+            as |curve|}}
+            {{capitalize curve}}
+          {{/power-select}}
+        </div>
+        <div class="line-chart-config__area-opt">
+          <span class="line-chart-config__area-opt-title">Area</span>
+          {{x-toggle
+            class="line-chart-config__area-opt-select"
+            value=(readonly options.style.area)
+            size="small"
+            onToggle=(action "onUpdateStyle" "area")
+          }}
+        </div>
+      {{/if}}
+      {{#if showStackOption}}
+        <div class="line-chart-config__stacked-opt">
+          <span class="line-chart-config__stacked-opt-title">Stacked</span>
+          {{x-toggle
+            class="line-chart-config__stacked-opt-select"
+            value=(readonly options.style.stacked)
+            size="small"
+            onToggle=(action "onUpdateStyle" "stacked")
+          }}
+        </div>
+      {{/if}}
+    </div>
+    {{#if (and showStackOption options.style.stacked)}}
+      {{chart-series-config
+        seriesConfig=seriesConfig
+        seriesType=seriesType
+        onUpdateConfig=(action "onUpdateConfig")
+        class="line-chart-config__series-config"
+      }}
+    {{/if}}
+  {{else if (not (eq seriesType "dimension"))}}
+    <div class="visualization-config__warn-msg">No configuration options available.</div>
+  {{/if}}
+{{/component}}
