+++ packages/core/addon/templates/components/visualization-config/pie-chart.hbs
@@ -1,11 +1,12 @@
 {{!-- Copyright 2019, Yahoo Holdings Inc. Licensed under the terms of the MIT license. See accompanying LICENSE.md file for terms. --}}
-{{#if (not (eq seriesType "dimension"))}}
+{{#if (not (eq options.series.type "dimension"))}}
   <div class="visualization-config__warn-msg">No configuration options available.</div>
 {{/if}}
 {{component
-  (concat typePrefix seriesType)
+  "visualization-config/series-chart"
   request=request
   response=response
   seriesConfig=(readonly options.series.config)
+  seriesType=(readonly options.series.type)
   onUpdateConfig=(action "onUpdateConfig")
 }}
