+++ packages/core/addon/templates/components/visualization-config/series-chart.hbs
@@ -1,0 +1,32 @@
+{{!-- Copyright 2019, Yahoo Holdings Inc. Licensed under the terms of the MIT license. See accompanying LICENSE.md file for terms. --}}
+{{#if showMetricSelect}}
+  <div class="metric-select">
+    <div class="metric-select__header">Metric</div>
+    <div class="metric-select__select">
+      {{#power-select
+        options=metrics
+        selected=selectedMetric
+        searchEnabled=false
+        onchange=(action "onUpdateChartMetric")
+        dropdownClass="metric-select__select__dropdown"
+        triggerClass="metric-select__select__trigger"
+        class="metric-select__select__selector"
+        tagName="div"
+        renderInPlace=true
+        as |metric|
+      }}
+        {{metric-format metric}}
+      {{/power-select}}
+    </div>
+  </div>
+{{/if}}
+{{yield}}
+{{#if (eq seriesType "dimension")}}
+  {{chart-series-collection
+      allSeriesData=allSeriesData
+      selectedSeriesData=selectedSeriesData
+      seriesDimensions=dimensions
+      onRemoveSeries=(action "onRemoveSeries")
+      onAddSeries=(action "onAddSeries")
+  }}
+{{/if}}
