--- packages/core/app/components/visualization-config/chart-type/date-time.js
@@ -1,1 +1,0 @@
-export { default } from 'navi-core/components/visualization-config/chart-type/date-time';
