--- packages/core/app/components/visualization-config/chart-type/dimension.js
@@ -1,1 +1,0 @@
-export { default } from 'navi-core/components/visualization-config/chart-type/dimension';
