+++ packages/core/app/components/visualization-config/series-chart.js
@@ -1,0 +1,1 @@
+export { default } from 'navi-core/components/visualization-config/series-chart';
