--- packages/core/app/styles/navi-core/components/visualization-config/chart-type/index.less
@@ -1,6 +1,0 @@
-/**
- * Copyright 2019, Yahoo Holdings Inc.
- * Licensed under the terms of the MIT license. See accompanying LICENSE.md file for terms.
- */
-
-@import 'metric';
