--- packages/core/app/styles/navi-core/components/visualization-config/chart-type/metric.less
@@ -1,52 +1,0 @@
-/**
- * Copyright 2017, Yahoo Holdings Inc.
- * Licensed under the terms of the MIT license. See accompanying LICENSE.md file for terms.
- */
-.metric-line-chart-config {
-  .ember-power-select-trigger {
-    border: none;
-    padding: 0;
-
-    &[aria-disabled='true'] {
-      opacity: 0.5;
-      background-color: @navi-white;
-      .btn.btn-add {
-        cursor: not-allowed;
-      }
-    }
-
-    .btn.btn-add {
-      background-color: @navi-blue;
-      color: @navi-white;
-      width: 100%;
-    }
-  }
-
-  .ember-power-select {
-    margin-bottom: 15px;
-  }
-
-  .ember-power-select-dropdown {
-    .generate-chart-colors-for(border-top-color);
-    border-top-style: solid;
-    border-top-width: 4px;
-    border-radius: 5px;
-    margin-top: 4px;
-
-    input {
-      border: none;
-      font-size: 14px;
-      padding-left: 12px;
-    }
-  }
-
-  .navi-icon__delete {
-    color: @navi-gray;
-    float: right;
-    cursor: pointer;
-
-    &:hover {
-      color: @navi-black;
-    }
-  }
-}
