+++ packages/core/app/styles/navi-core/components/visualization-config/index.less
@@ -3,9 +3,9 @@
  * Licensed under the terms of the MIT license. See accompanying LICENSE.md file for terms.
  */
 
+@import 'series-chart';
 @import 'line-chart';
 @import 'wrapper';
-@import 'chart-type/index';
 @import 'goal-gauge';
 @import 'metric-label';
 @import 'pie-chart';
